requirements.txt
pdfrw
google-cloud-storage

